Add animation to Game Intro page
