+++ app/src/main/res/animator/scale.xml
@@ -3,11 +3,11 @@
 
     <objectAnimator xmlns:android="http://schemas.android.com/apk/res/android"
         android:duration="500"
-        android:propertyName="translationX"
+        android:propertyName="alpha"
         android:repeatCount="1"
         android:repeatMode="reverse"
-        android:valueFrom="0"
-        android:valueTo="200"
+        android:valueFrom="1.0"
+        android:valueTo="0.0"
         android:valueType="floatType" />
 
 </set>
